backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-genai>=1.0.0
httpx>=0.26.0
pydantic>=2.5.0
Pillow>=10.0.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
